refactor: 將 odometry 計算邏輯改為使用 KinematicModel 類別
- 新增 VehicleHardwardData 結構為 MeamrBase 的成員變數
- 使用 kinematic_model_.forward_kinematics() 替代原本硬編碼的 odometry 計算公式

+++ include/meamr_drive_model/dd_kinematic_model.hpp
@@ -6,7 +6,7 @@ class KinematicModel
 {
 public:
     KinematicModel();
-    void forward_kinematics(float leftMotorRPM, float rightMotorRPM, const VehicleHardwardData &data, double &outLinearVel, double &outAngularVel);
+    void forward_kinematics(double leftMotorRPM, double rightMotorRPM, const VehicleHardwardData &data, double &outLinearVel, double &outAngularVel);
     void backward_kinematics(double linearVel, double angularVel, const VehicleHardwardData &data, float &outLeftMotorRPM, float &outRightMotorRPM);
 };
